extract some methods
